+++ src/app/preview/[id]/page.tsx
@@ -4,9 +4,7 @@ import { useEffect, useState } from "react";
 import { useParams } from "next/navigation";
 import { ExternalLinkIcon } from "lucide-react";
 import { ChevronRightIcon } from "lucide-react";
-import { Share2 } from "lucide-react";
-import { Button } from "@/components/ui/button";
-
+import {ShareButton} from  "../../../components/common/ShareButton";
 
 interface WebsiteData {
     websiteName: string;
@@ -357,65 +355,4 @@ className="text-sm text-blue-600 hover:underline pl-6"
     );
 
 
-}
-//exporting share options
-import {
-  FacebookShareButton,
-  WhatsappShareButton,
-  TwitterShareButton,
-  FacebookIcon,
-  WhatsappIcon,
-  TwitterIcon,
-} from "next-share";
-
-export function SocialShare({ url, title }: { url: string; title: string }) {
-  return (
-    <div className="flex gap-2">
-      <FacebookShareButton url={url} quote={title}>
-        <FacebookIcon size={32} round />
-      </FacebookShareButton>
-      <WhatsappShareButton url={url} title={title}>
-        <WhatsappIcon size={32} round />
-      </WhatsappShareButton>
-      <TwitterShareButton url={url} title={title}>
-        <TwitterIcon size={32} round />
-      </TwitterShareButton>
-    </div>
-  );
-}
-
-
-export function ShareButton({ title, url }: { title: string; url: string }) {
-  const handleShare = async () => {
-    if (navigator.share) {
-      try {
-        await navigator.share({
-          title,
-          text: `Check this out: ${title}`,
-          url,
-        });
-      } catch (err) {
-        //console.error("Share failed:", err);
-      }
-    } else {
-      // fallback: copy link to clipboard
-      try {
-        await navigator.clipboard.writeText(url);
-        alert("🔗 Link copied to clipboard!");
-      } catch (err) {
-        console.error("Clipboard write failed:", err);
-      }
-    }
-  };
-
-  return (
-    <Button
-      variant="outline"
-      size="icon"
-      onClick={handleShare}
-      className="flex items-center"
-    >
-      <Share2 className="h-4 w-4" />
-    </Button>
-  );
 }
